Fix implementation
Previous commit was not a finished implementation of the decoder. It did
not work and had numerous issues. This push finishes the decoder.

This is based on the reference implementation of QOIF, so hats off to
them for their implementation. It was better than my first go at it, so
I just rebased on theirs (but in Rust).