Switch to wrapping operations in decode
A previous commit accidentally included wrapping changes to hash_pixel,
so this commit is finalizing the last wrapping code.

Previously, the code only worked on release (or if you manually disabled
wrapping checks. The standard is designed around wrapping u8s, so this
is required.

On a side note, Rust code that needs to wrap is gross and horrible to
write. How has this stayed this bad for this long (the Wrapping type
came out like at 1.12 or something)