Switch to bin/lib cargo project
This project is meant to be both a binary and a library. This change
helps to get there by moving main.rs to bin/qoi-parser.rs. All other
code remains as is, and a new lib.rs file imports all of it as public.
The Cargo.toml file was also updated to match this new structure.