Add streaming decoder
Big bulk commit that adds StreamDecoder, which is a byte-by-byte
streaming decoder for the QOI standard. This allows for a user to parse
a QOI image without having to load the entire image into memory. Each
byte can be passed into the parser and individual resulting pixels can
be handled as they appear.

This is implemented as a big state machine operating on a struct. Users
feed bytes in using the StreamDecoder::feed() method, and results of
type StreamDecoderOutput are returned. These results inform the user of
whether there are pixels available, more data is need, or if the image
is finished. All outputs are wrapped in a Result, allowing for error
propagation while parsing.

This implementation is verbose, probably slow, and needs some cleanup.
I'd like to go through and abstract out to methods to cleanup the state
machine. This also probably uses more memory than it needs to. Also, the
dec.rs file is nearly 1000 lines long. It need to be broken out into
more files.