Move stream decoder to new module
The dec.rs file was getting a bit dense, so this moves the stream
decoder out to a stream module. There's still some work that needs to be
done here, mostly with weird connections between dec.rs and
stream/dec.rs, but those are small and can be fixed later.